refactor: Component 패키지 분리
