+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -21,9 +21,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.signup.ui.ButtonComponent
-import nextstep.signup.ui.TextComponent
-import nextstep.signup.ui.TextFieldComponent
+import nextstep.signup.ui.component.ButtonComponent
+import nextstep.signup.ui.component.TextComponent
+import nextstep.signup.ui.component.TextFieldComponent
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
