+++ app/src/main/java/nextstep/signup/ui/component/ButtonComponent.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui
+package nextstep.signup.ui.component
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.PaddingValues
